refactor config code to be a helper
